Added missing demote to plugin.yml
attempted fix in database lockup...

Im 90% sure theres a typo by the author, where if its a locking area its meant to brake out the loop, but they wrote false instead of true.

however, rather than bailing in this case, i have allowed it to try and run the quiery a maximum of 10 times before bailing.
should increase chances of query being succesful, but not result in a lockup :D maybe?

git-svn-id: https://subversion.assembla.com/svn/wisbycraft-perks@52 2df59c3f-94b9-4873-9e60-c579250e45c4