Nuked magkilled magic carpet. added extra return on player move...
The slowest part about perks is the finding of a player...
say there are 30 players, and the 30th player is afk. it means we have to check the other 29 players before we find the player we want.
It might be worth sorting the list of players...
could change it based upon /afk status ect...


git-svn-id: https://subversion.assembla.com/svn/wisbycraft-perks@108 2df59c3f-94b9-4873-9e60-c579250e45c4